+++ src/onegov.event/onegov/event/utils/__init__.py
@@ -1,6 +1,8 @@
 from onegov.event.utils.guidle import GuidleExportData
+from onegov.event.utils.utils import as_rdates
 
 
 __all__ = (
+    'as_rdates',
     'GuidleExportData',
 )